jobmanager: Setup pubsub to all web instances for 'jobresult'
A bit hacky, code requires information both about the web frontend
participant, and the graph topology. And only works for things
inbound to the web process.
Eventually we want a cleaned-up version of this in MsgFlo itself,
but it seems it will require a continiously running coordinator
in the general case...

Should allow u